+++ notification-infrastructure/src/main/kotlin/io/github/v1servicenotification/domain/setting/facade/NotificationSettingFacade.kt
@@ -3,7 +3,7 @@ package io.github.v1servicenotification.domain.setting.facade
 import io.github.v1servicenotification.domain.category.facade.CategoryFacade
 import io.github.v1servicenotification.domain.setting.domain.SettingEntity
 import io.github.v1servicenotification.domain.setting.domain.SettingId
-import io.github.v1servicenotification.domain.setting.domain.repository.NotificationSettingRepository
+import io.github.v1servicenotification.domain.setting.domain.repository.SettingRepository
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Component
 import java.util.*
@@ -11,7 +11,7 @@ import java.util.*
 @Component
 class NotificationSettingFacade(
     private val categoryFacade: CategoryFacade,
-    private val notificationSettingRepository: NotificationSettingRepository
+    private val settingRepository: SettingRepository
 ) {
 
     fun saveOrUpdateNotificationSetting(categoryUUID: UUID, isActivate: Boolean): Int {
@@ -22,12 +22,12 @@ class NotificationSettingFacade(
             categoryEntity = notificationCategory
         )
 
-        return if (notificationSettingRepository.findBySettingId(settingId) != null) {
-            notificationSettingRepository.findBySettingId(settingId)!!
+        return if (settingRepository.findBySettingId(settingId) != null) {
+            settingRepository.findBySettingId(settingId)!!
                 .changeIsActivate(isActivate)
             HttpStatus.NO_CONTENT.value()
         } else {
-            notificationSettingRepository.save(
+            settingRepository.save(
                 SettingEntity(
                     settingId = settingId,
                     isActivated = isActivate
